Move CTeam traits
